+++ src/untangled/server/fixtures/dont_require_me.clj
@@ -1,0 +1,4 @@
+(ns untangled.server.fixtures.dont-require-me
+  "For use in untangled.server.impl.components.config-spec ONLY")
+
+(def stahp :stop-it)