+++ src/untangled/server/impl/util.clj
@@ -1,0 +1,16 @@
+(ns untangled.server.impl.util
+  (:require
+    om.tempid
+    [clojure.tools.namespace.find :refer [find-namespaces]]
+    [clojure.java.classpath :refer [classpath]]
+    [om.next.impl.parser :as omp]))
+
+(defn deep-merge [& xs]
+  (if (every? map? xs)
+    (apply merge-with deep-merge xs)
+    (last xs)))
+
+(defn strip-parameters
+  "Removes parameters from the query, e.g. for PCI compliant logging."
+  [query]
+  (-> (clojure.walk/prewalk #(if (map? %) (dissoc % :params) %) (omp/query->ast query)) (omp/ast->expr true)))